# docs/superpowers/specs/2026-04-21-deck-go-control-plane-api-openapi.yaml
openapi: 3.1.0
info:
  title: OpenClaw Deck Control Plane API
  version: 0.1.0-stage2
  summary: Stage 2 target query/command HTTP contract for the control plane
servers:
  - url: /api/v1
paths:
...
  /webhooks/{webhookId}:
    patch:
      summary: Update a webhook
      operationId: updateWebhook
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Webhook update payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    delete:
      summary: Delete a webhook
      operationId: deleteWebhook
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Webhook delete payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /usage/budget:
    get:
      summary: Fetch budget rules
      operationId: listBudgetRules
      responses:
        "200":
          description: Budget rule collection payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    post:
      summary: Create a budget rule
      operationId: createBudgetRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Budget rule create payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /runtimes/{runtimeId}/runs:
    get:
      summary: List projected runs for a runtime
      operationId: listRuns
      parameters:
        - $ref: "#/components/parameters/RuntimeId"
      responses:
        "200":
          description: Run collection
          content:
            application/json:
              schema:
                type: object
                properties:
                  runtimeId:
                    type: string
                  runs:
                    type: array
                    items:
                      $ref: "#/components/schemas/RunSummary"
                  requestId:
                    type: string
                required: [runtimeId, runs, requestId]
        "404":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    RuntimeId:
      name: runtimeId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Standard control-plane error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    RunSummary:
      type: object
      properties:
        runId:
          type: string
        agentId:
          type: string
          nullable: true
        sessionKey:
          type: string
          nullable: true
        status:
          type: string
        firstEventAt:
          type: string
        lastEventAt:
          type: string
        eventCount:
          type: integer
        toolCalls:
          type: integer
        modelCalls:
          type: integer
        fileOps:
          type: integer
        subagentSpawns:
          type: integer
        totalTokens:
          type: integer
        compacted:
          type: boolean
      required: [runId, status, firstEventAt, lastEventAt, eventCount]
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true
          required: [code, message, details]
        requestId:
          type: string
      required: [error, requestId]